+++ python/pytauri/src/pytauri/webview.py
@@ -1,0 +1,5 @@
+"""[tauri::webview](https://docs.rs/tauri/latest/tauri/webview/index.html)"""
+
+from pytauri.ffi.webview import WebviewWindow
+
+__all__ = ["WebviewWindow"]