+++ python/pytauri/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- [#46](https://github.com/WSH032/pytauri/pull/46) - add `webview::WebviewWindow`, `Manager`, `ImplManager`, `App::handle`
+
 ## [0.1.0-beta.0]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD